feat(测试计划): 初始化规划节点
