+++ backend/framework/domain/src/main/java/io/metersphere/plan/domain/TestPlanCollection.java
@@ -24,20 +24,29 @@ public class TestPlanCollection implements Serializable {
     @Size(min = 1, max = 50, message = "{test_plan_collection.test_plan_id.length_range}", groups = {Created.class, Updated.class})
     private String testPlanId;
 
-    @Schema(description = "所属测试集类型ID", requiredMode = Schema.RequiredMode.REQUIRED)
-    @NotBlank(message = "{test_plan_collection.test_collection_type_id.not_blank}", groups = {Created.class})
-    @Size(min = 1, max = 50, message = "{test_plan_collection.test_collection_type_id.length_range}", groups = {Created.class, Updated.class})
-    private String testCollectionTypeId;
+    @Schema(description = "父级ID", requiredMode = Schema.RequiredMode.REQUIRED)
+    @NotBlank(message = "{test_plan_collection.parent_id.not_blank}", groups = {Created.class})
+    @Size(min = 1, max = 50, message = "{test_plan_collection.parent_id.length_range}", groups = {Created.class, Updated.class})
+    private String parentId;
 
     @Schema(description = "测试集名称", requiredMode = Schema.RequiredMode.REQUIRED)
     @NotBlank(message = "{test_plan_collection.name.not_blank}", groups = {Created.class})
     @Size(min = 1, max = 255, message = "{test_plan_collection.name.length_range}", groups = {Created.class, Updated.class})
     private String name;
 
+    @Schema(description = "测试集类型(功能/接口/场景)", requiredMode = Schema.RequiredMode.REQUIRED)
+    @NotBlank(message = "{test_plan_collection.type.not_blank}", groups = {Created.class})
+    @Size(min = 1, max = 255, message = "{test_plan_collection.type.length_range}", groups = {Created.class, Updated.class})
+    private String type;
+
     @Schema(description = "执行方式(串行/并行)", requiredMode = Schema.RequiredMode.REQUIRED)
     @NotBlank(message = "{test_plan_collection.execute_method.not_blank}", groups = {Created.class})
     @Size(min = 1, max = 50, message = "{test_plan_collection.execute_method.length_range}", groups = {Created.class, Updated.class})
     private String executeMethod;
+
+    @Schema(description = "是否继承", requiredMode = Schema.RequiredMode.REQUIRED)
+    @NotNull(message = "{test_plan_collection.extended.not_blank}", groups = {Created.class})
+    private Boolean extended;
 
     @Schema(description = "是否使用环境组", requiredMode = Schema.RequiredMode.REQUIRED)
     @NotNull(message = "{test_plan_collection.grouped.not_blank}", groups = {Created.class})
@@ -48,9 +57,31 @@ public class TestPlanCollection implements Serializable {
     @Size(min = 1, max = 50, message = "{test_plan_collection.environment_id.length_range}", groups = {Created.class, Updated.class})
     private String environmentId;
 
-    @Schema(description = "自定义排序，间隔为2的N次幂", requiredMode = Schema.RequiredMode.REQUIRED)
-    @NotNull(message = "{test_plan_collection.pos.not_blank}", groups = {Created.class})
-    private Long pos;
+    @Schema(description = "测试资源池ID", requiredMode = Schema.RequiredMode.REQUIRED)
+    @NotBlank(message = "{test_plan_collection.test_resource_pool_id.not_blank}", groups = {Created.class})
+    @Size(min = 1, max = 50, message = "{test_plan_collection.test_resource_pool_id.length_range}", groups = {Created.class, Updated.class})
+    private String testResourcePoolId;
+
+    @Schema(description = "是否失败重试", requiredMode = Schema.RequiredMode.REQUIRED)
+    @NotNull(message = "{test_plan_collection.retry_on_fail.not_blank}", groups = {Created.class})
+    private Boolean retryOnFail;
+
+    @Schema(description = "失败重试类型(步骤/场景)", requiredMode = Schema.RequiredMode.REQUIRED)
+    @NotBlank(message = "{test_plan_collection.retry_type.not_blank}", groups = {Created.class})
+    @Size(min = 1, max = 50, message = "{test_plan_collection.retry_type.length_range}", groups = {Created.class, Updated.class})
+    private String retryType;
+
+    @Schema(description = "失败重试次数", requiredMode = Schema.RequiredMode.REQUIRED)
+    @NotNull(message = "{test_plan_collection.retry_times.not_blank}", groups = {Created.class})
+    private Integer retryTimes;
+
+    @Schema(description = "失败重试间隔(单位: ms)", requiredMode = Schema.RequiredMode.REQUIRED)
+    @NotNull(message = "{test_plan_collection.retry_interval.not_blank}", groups = {Created.class})
+    private Integer retryInterval;
+
+    @Schema(description = "是否失败停止", requiredMode = Schema.RequiredMode.REQUIRED)
+    @NotNull(message = "{test_plan_collection.stop_on_fail.not_blank}", groups = {Created.class})
+    private Boolean stopOnFail;
 
     @Schema(description = "创建人")
     private String createUser;
@@ -58,19 +89,31 @@ public class TestPlanCollection implements Serializable {
     @Schema(description = "创建时间")
     private Long createTime;
 
+    @Schema(description = "自定义排序，间隔为4096", requiredMode = Schema.RequiredMode.REQUIRED)
+    @NotNull(message = "{test_plan_collection.pos.not_blank}", groups = {Created.class})
+    private Long pos;
+
     private static final long serialVersionUID = 1L;
 
     public enum Column {
         id("id", "id", "VARCHAR", false),
         testPlanId("test_plan_id", "testPlanId", "VARCHAR", false),
-        testCollectionTypeId("test_collection_type_id", "testCollectionTypeId", "VARCHAR", false),
+        parentId("parent_id", "parentId", "VARCHAR", false),
         name("name", "name", "VARCHAR", true),
+        type("type", "type", "VARCHAR", true),
         executeMethod("execute_method", "executeMethod", "VARCHAR", false),
+        extended("extended", "extended", "BIT", false),
         grouped("grouped", "grouped", "BIT", false),
         environmentId("environment_id", "environmentId", "VARCHAR", false),
-        pos("pos", "pos", "BIGINT", false),
+        testResourcePoolId("test_resource_pool_id", "testResourcePoolId", "VARCHAR", false),
+        retryOnFail("retry_on_fail", "retryOnFail", "BIT", false),
+        retryType("retry_type", "retryType", "VARCHAR", false),
+        retryTimes("retry_times", "retryTimes", "INTEGER", false),
+        retryInterval("retry_interval", "retryInterval", "INTEGER", false),
+        stopOnFail("stop_on_fail", "stopOnFail", "BIT", false),
         createUser("create_user", "createUser", "VARCHAR", false),
-        createTime("create_time", "createTime", "BIGINT", false);
+        createTime("create_time", "createTime", "BIGINT", false),
+        pos("pos", "pos", "BIGINT", false);
 
         private static final String BEGINNING_DELIMITER = "`";
