+++ backend/framework/sdk/src/main/resources/i18n/plan.properties
@@ -113,4 +113,6 @@ test_plan.is.archived=测试计划已归档
 test_plan.cannot.archived=测试计划不符合归档操作条件
 test_plan_module_already_exists=同名模块已存在
 test_plan_report_name_length_range=报告名称长度过长
-test_plan_allocation_type_param_error=测试集所属分类参数错误
+test_plan_default_functional_collection_name=基本功能点
+test_plan_default_api_collection_name=单接口验证
+test_plan_default_scenario_collection_name=业务流程验证
