+++ backend/framework/sdk/src/main/resources/i18n/plan_en_US.properties
@@ -116,4 +116,6 @@ test_plan.is.archived=Test plan has been archived
 test_plan.cannot.archived=Test plan cannot be archived
 test_plan_module_already_exists=The module with the same name already exists
 test_plan_report_name_length_range=The report name is too long
-test_plan_allocation_type_param_error=The parameter of the allocation type is not correct
+test_plan_default_functional_collection_name=Basic function point
+test_plan_default_api_collection_name=Single interface verification
+test_plan_default_scenario_collection_name=Business process verification
