+++ backend/framework/sdk/src/main/resources/i18n/plan_zh_TW.properties
@@ -116,3 +116,6 @@ test_plan.cannot.archived=測試計劃不符合歸檔操作條件
 test_plan_module_already_exists=同名模塊已存在
 test_plan_report_name_length_range=报告名称长度过长
 test_plan_allocation_type_param_error=測試集所屬分類參數錯誤
+test_plan_default_functional_collection_name=基本功能點
+test_plan_default_api_collection_name=單接口驗證
+test_plan_default_scenario_collection_name=業務流程驗證
