+++ backend/services/test-plan/src/main/java/io/metersphere/plan/enums/RetryType.java
@@ -1,0 +1,14 @@
+package io.metersphere.plan.enums;
+
+public enum RetryType {
+
+	/**
+	 * 步骤
+	 */
+	STEP,
+
+	/**
+	 * 场景
+	 */
+	SCENARIO
+}
