+++ backend/services/test-plan/src/test/resources/dml/init_test_plan_test.sql
@@ -50,11 +50,6 @@ VALUES
     ('wx_test_plan_id_7', b'0', b'0', 100);
 
 
-
-INSERT INTO `test_plan_allocation`(`id`, `test_plan_id`, `test_resource_pool_id`, `retry_on_fail`, `retry_type`, `retry_times`, `retry_interval`, `stop_on_fail`)
-VALUES ('1', 'wx_test_plan_id_1', '111', b'0', 'scenario', '10', '1000', b'0');
-
-
 INSERT INTO functional_case(id, num, module_id, project_id, template_id, name, review_status, tags, case_edit_type, pos, version_id, ref_id, last_execute_result, deleted, public_case, latest, create_user, update_user, delete_user, create_time, update_time, delete_time)
 VALUES ('my_test_1', 1, '1', 'songtianyang-fix-wx', '100001', '1111', 'UN_REVIEWED', NULL, 'TEXT', 55000, 'v3.0.0',
         'TEST_FUNCTIONAL_MINDER_CASE_ID_7', 'UN_EXECUTED', b'0', b'0', b'1', 'admin', 'admin', '', 1698058347559,
@@ -65,4 +60,9 @@ VALUES ('my_test_1', 1, '1', 'songtianyang-fix-wx', '100001', '1111', 'UN_REVIEW
        ('my_test_3', 3, 'root', 'songtianyang-fix-wx', '100001', '3333', 'UN_REVIEWED', NULL, 'TEXT', 55000, 'v3.0.0',
         'TEST_FUNCTIONAL_MINDER_CASE_ID_7', 'UN_EXECUTED', b'0', b'0', b'1', 'admin', 'admin', '', 1698058347559,
         1698058347559, NULL);
+
+-- 初始化计划默认节点
+INSERT INTO `test_plan`(`id`, `num`, `project_id`, `group_id`, `module_id`, `name`, `status`, `type`, `tags`, `create_time`, `create_user`, `update_time`, `update_user`, `planned_start_time`, `planned_end_time`, `actual_start_time`, `actual_end_time`, `description`)
+VALUES ('init_plan_id', 5000, 'init_project', 'NONE', '1', '测试一下计划', 'PREPARED', 'TEST_PLAN', NULL,
+        unix_timestamp() * 1000, 'admin', unix_timestamp() * 1000, 'admin', unix_timestamp() * 1000, unix_timestamp() * 1000, unix_timestamp() * 1000, unix_timestamp() * 1000, '11');
